+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/engine/EngineObserver.kt
@@ -443,6 +443,10 @@ internal class EngineObserver(
         store.dispatch(EngineAction.PrintContentExceptionAction(tabId, isPrint, throwable))
     }
 
+    override fun onSaveToPdfComplete() {
+        store.dispatch(EngineAction.SaveToPdfCompleteAction(tabId))
+    }
+
     override fun onCheckForFormData(containsFormData: Boolean) {
         store.dispatch(ContentAction.UpdateHasFormDataAction(tabId, containsFormData))
     }