+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/reducer/EngineStateReducer.kt
@@ -45,7 +45,9 @@ internal object EngineStateReducer {
         is EngineAction.OptimizedLoadUrlTriggeredAction -> {
             state
         }
-        is EngineAction.SaveToPdfExceptionAction -> {
+        is EngineAction.SaveToPdfExceptionAction,
+        is EngineAction.SaveToPdfCompleteAction,
+        -> {
             throw IllegalStateException(
                 "You need to add a middleware to handle this action in your BrowserStore. ($action)",
             )