+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/share/ShareController.kt
@@ -146,7 +146,6 @@ class DefaultShareController(
     }
 
     override fun handleSaveToPDF(tabId: String?) {
-        Events.saveToPdfTapped.record(NoExtras())
         handleShareClosed()
         saveToPdfUseCase.invoke(tabId)
     }